backend/requirements.txt
flask==3.0.3
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1 